requirements.txt
python-dotenv
flask-sqlalchemy
flask-cors